requirements.txt
PyVCF==0.4.4
